requirements.txt
playwright
beautifulsoup4
chromadb
uuid
google-generativeai
python-dotenv
